feat: actualizaciones pendientes en staging

+++ src/main/java/com/apso/app/service/CargaCSVService.java
@@ -41,7 +41,8 @@ public class CargaCSVService {
                 estudiante.setGrupoTeorico(record.get("grupo_teorico"));
                 estudiante.setAsignatura(record.get("asignatura"));
                 //estudiante.setCargaId(record.get("carga_id"));
-                estudiante.setCargaId(Integer.parseInt(record.get("carga_id")));
+                estudiante.setCargaId(Long.parseLong(record.get("carga_id")));
+
 
                 estudiantes.add(estudiante);
             }
